refactor(Options): move each settings page to its own subdir
