+++ src/pages/options/Donate/index.tsx
@@ -4,8 +4,8 @@ import Typography from '@mui/material/Typography';
 import { usePageTitle } from '@root/src/shared/contexts/PageTitleContext';
 import Rolod0xText from '@src/components/Rolod0xText';
 
-import SettingsPageHeader from './shared/SettingsPageHeader';
-import StyledCode from './shared/StyledCode';
+import SettingsPageHeader from '../shared/SettingsPageHeader';
+import StyledCode from '../shared/StyledCode';
 
 export default function Donate() {
   const { setPageTitle } = usePageTitle();
